Made camera follow Spline
I made the camera follow a spline. This took a while and I rewrote most of the code to make the camera follow the player. The new system has a tendency to settle looking down at the player which needs to be fixed. Might make it so the camera only moves along the spline if it's a certain distance from the player later.